Move TagPluginChunk and implement RegisterPlugin
